allow m.nviz.image (and maybe the rest of the nviz family) to work with 64bit region sizes (merge from trunk)
git-svn-id: https://svn.osgeo.org/grass/grass/branches/develbranch_6@58787 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ include/ogsf_proto.h
@@ -404,7 +404,7 @@ int gs_num_datah_reused(int);
 int gs_get_att_type(geosurf *, int);
 int gs_get_att_src(geosurf *, int);
 typbuff *gs_get_att_typbuff(geosurf *, int, int);
-int gs_malloc_att_buff(geosurf *, int, int);
+size_t gs_malloc_att_buff(geosurf *, int, int);
 int gs_malloc_lookup(geosurf *, int);
 int gs_set_att_type(geosurf *, int, int);
 int gs_set_att_src(geosurf *, int, int);
